alpaca_bot/requirements_ui.txt
# Complete requirements for Alpaca Trading Bot with UI
alpaca-py==0.43.1
python-dotenv==1.0.0
pandas==2.0.3
requests==2.31.0

# UI Dependencies
streamlit==1.41.1
plotly==6.0.0
streamlit-autorefresh==1.0.1

# Optional: Enhanced features
schedule==1.2.0